Convert optimistic challenges hooks into selector (#872)
* Move hooks to selectors

* Remove hooks

* Rearrange to make diff less confusing

* Add docs back
