+++ src/components/nav/desktop/NavAudio.tsx
@@ -8,13 +8,13 @@ import { ReactComponent as IconCaretRight } from 'assets/img/iconCaretRight.svg'
 import IconNoTierBadge from 'assets/img/tokenBadgeNoTier.png'
 import { FeatureFlags } from 'common/services/remote-config/feature-flags'
 import { getAccountUser } from 'common/store/account/selectors'
+import { getOptimisticUserChallenges } from 'common/store/challenges/selectors/optimistic-challenges'
 import { getAccountTotalBalance } from 'common/store/wallet/selectors'
 import { formatWei } from 'common/utils/wallet'
 import { audioTierMapPng } from 'components/user-badges/UserBadges'
 import { useSelectTierInfo } from 'components/user-badges/hooks'
 import { useNavigateToPage } from 'hooks/useNavigateToPage'
 import { useFlag } from 'hooks/useRemoteConfig'
-import { useOptimisticUserChallenges } from 'pages/audio-rewards-page/hooks'
 import { useSelector } from 'utils/reducer'
 import { AUDIO_PAGE } from 'utils/route'
 
@@ -43,7 +43,7 @@ const NavAudio = () => {
   const { tier } = useSelectTierInfo(account?.user_id ?? 0)
   const audioBadge = audioTierMapPng[tier]
 
-  const userChallenges = useOptimisticUserChallenges()
+  const userChallenges = useSelector(getOptimisticUserChallenges)
   const hasExpiredClaim = Object.values(userChallenges).some(
     challenge => challenge?.state === 'completed'
   )
